refactor openai to package
